requirements.txt
torch
torchvision
Pillow
numpy
scikit-image
tqdm
matplotlib
scikit-learn